pruebaMonoku/requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
Faker==0.9.1
flake8==3.9.0
gunicorn==20.1.0
importlib-metadata==4.0.1
mccabe==0.6.1
numpy==1.20.2
pandas==1.2.4
psycopg2==2.8.5
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.2
typing-extensions==3.7.4.3
whitenoise==5.2.0
zipp==3.4.1
